supabase/backEnd/db/csv/requirements_csv.txt
supabase==2.9.0
python-dotenv==1.0.0
httpx==0.27.0
websockets>=13.0
